refactored files done with ui
